requirements.txt
Django==4.1.3
django-ninja~=0.19.1